Fix: git merge 과정에서 사라진 코드 재생성, 잡다한 버그 수정
리포지토리,  totalPrice 계산 코드, 빌더 파라미터

+++ src/main/java/org/kakao/kakaoshopping/web/dto/order/response/OrderItemSimpleView.java
@@ -20,7 +20,7 @@ public class OrderItemSimpleView {
 	private Long itemId;
 	private String itemName;
 	private Integer quantity;
-	private BigDecimal price;
+	private Integer price;
 	private String imagePath;
 
 	public OrderItemSimpleView(OrderItem orderItem) {
